phase-2/backend/requirements.txt
# FastAPI Backend Dependencies
# Feature: Authentication (F001)

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation
pydantic[email]==2.5.3
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
python-dotenv==1.0.0
